move to rnafusion model
