+++ cg/models/nf_analysis.py
@@ -1,19 +1,9 @@
-@classmethod
-def statuses(cls) -> Tuple:
-    return tuple(status.value for status in cls)
-
-
-import collections
 from pathlib import Path
-from typing import List, Optional, Union
+from typing import List, Optional
 
-from pydantic.v1 import BaseModel, FilePath, conlist, root_validator, validator
+from pydantic.v1 import BaseModel, FilePath, conlist, validator
 
-from cg import resources
-from cg.constants.constants import FileFormat
-from cg.constants.nextflow import DELIVER_FILE_HEADERS
 from cg.exc import SampleSheetError
-from cg.io.controller import ReadFile
 
 
 class PipelineParameters(BaseModel):
@@ -65,29 +55,3 @@ class PipelineDeliverables(BaseModel):
     """Specification for pipeline deliverables."""
 
     files: List[FileDeliverable]
-
-
-class RnafusionDeliverables(PipelineDeliverables):
-    """Specification for pipeline deliverables."""
-
-    @staticmethod
-    def get_deliverables_template() -> List[dict]:
-        """Return deliverables file template content."""
-        return ReadFile.get_content_from_file(
-            file_format=FileFormat.YAML,
-            file_path=resources.RNAFUSION_BUNDLE_FILENAMES_PATH,
-        )
-
-    @classmethod
-    def get_deliverables_for_case(cls, case_id: str, case_path: Path):
-        """Return RnafusionDeliverables for a given case."""
-        template: List[dict] = cls.get_deliverables_template()
-        files: list = []
-        for file in template:
-            for key, value in file.items():
-                if value is None:
-                    continue
-                file[key] = file[key].replace("CASEID", case_id)
-                file[key] = file[key].replace("PATHTOCASE", str(case_path))
-            files.append(FileDeliverable(**file))
-        return cls(files=files)
